+++ packages/elements/src/autosuggest/index.ts
@@ -46,10 +46,10 @@ export { queryWordSelect, itemRenderer, escapeRegExp, itemHighlightable, updateE
  * @attr {boolean} opened - Set to open auto suggest popup
  * @prop {boolean} [opened=false] -  Auto suggest popup's open state
  */
-@customElement('ef-autosuggest')
+@customElement('ef-autosuggest', {
+  alias: 'emerald-autosuggest'
+})
 export class Autosuggest extends Overlay {
-
-
   /**
    * A basic regexp matching pattern to replace text based on string input.
    * @param text Value to test against