+++ packages/elements/src/collapse/index.ts
@@ -21,10 +21,10 @@ import '../icon';
  * @slot header-right - Slot to add custom contents to the right side of header e.g. ef-icon, ef-checkbox
  *
  */
-@customElement('ef-collapse')
+@customElement('ef-collapse', {
+  alias: 'coral-collapse'
+})
 export class Collapse extends BasicElement {
-
-
   /**
    * A `CSSResult` that will be used
    * to style the host, slotted children