+++ packages/elements/src/sidebar-layout/index.ts
@@ -22,7 +22,9 @@ import '../layout';
  * @slot main-header - Main header.
  * @slot main-content - Content of main section.
  */
-@customElement('ef-sidebar-layout')
+@customElement('ef-sidebar-layout', {
+  alias: 'carbon-sidebar-layout'
+})
 export class SidebarLayout extends BasicElement {
   /**
    * A `CSSResult` that will be used