+++ packages/demo-block/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-alpha.6](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/demo-block@5.0.0-alpha.5...@refinitiv-ui/demo-block@5.0.0-alpha.6) (2021-06-11)
+
+**Note:** Version bump only for package @refinitiv-ui/demo-block
+
+
+
+
+
+# [5.0.0-alpha.5](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/demo-block@5.0.0-alpha.4...@refinitiv-ui/demo-block@5.0.0-alpha.5) (2021-06-11)
+
+**Note:** Version bump only for package @refinitiv-ui/demo-block
+
+
+
+
+
 # [5.0.0-alpha.4](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/demo-block@5.0.0-alpha.3...@refinitiv-ui/demo-block@5.0.0-alpha.4) (2021-06-09)
 
 **Note:** Version bump only for package @refinitiv-ui/demo-block