+++ packages/elements/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-alpha.8](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/elements@5.0.0-alpha.7...@refinitiv-ui/elements@5.0.0-alpha.8) (2021-06-11)
+
+
+### Features
+
+* remove deprecated methods and properties ([90d4513](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/commits/90d4513ebff7c1825fe3d638b787408b5d993598))
+* use element alias ([7295c1f](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/commits/7295c1fc6ed702ea8558f60b0b589ceecf38360b))
+
+
+
+
+
+# [5.0.0-alpha.7](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/elements@5.0.0-alpha.6...@refinitiv-ui/elements@5.0.0-alpha.7) (2021-06-11)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
+# [5.0.0-alpha.6](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/elements@5.0.0-alpha.5...@refinitiv-ui/elements@5.0.0-alpha.6) (2021-06-10)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
+# [5.0.0-alpha.5](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/elements@5.0.0-alpha.4...@refinitiv-ui/elements@5.0.0-alpha.5) (2021-06-10)
+
+**Note:** Version bump only for package @refinitiv-ui/elements
+
+
+
+
+
 # [5.0.0-alpha.4](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/elements@5.0.0-alpha.3...@refinitiv-ui/elements@5.0.0-alpha.4) (2021-06-09)
 
 **Note:** Version bump only for package @refinitiv-ui/elements